refactor: 서비스 이름 변경에 따라 linkpick으로 수정
