+++ backend/.env.sample
@@ -1,4 +1,4 @@
 DOCKER_MYSQL_USERNAME=root
 DOCKER_MYSQL_PASSWORD=1234
-DOCKER_MYSQL_DATABASE=TECHPICK_DB
-DOCKER_MYSQL_URL=jdbc:mysql://localhost:3306/TECHPICK_DB?createDatabaseIfNotExist=true
+DOCKER_MYSQL_DATABASE=LINKPICK_DB
+DOCKER_MYSQL_URL=jdbc:mysql://localhost:3306/LINKPICK_DB?createDatabaseIfNotExist=true
