+++ backend/src/main/java/kernel360/linkpick/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package kernel360.linkpick.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
