--- backend/src/main/java/kernel360/techpick/TechpickApplication.java
@@ -1,13 +1,0 @@
-package kernel360.techpick;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TechpickApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TechpickApplication.class, args);
-	}
-
-}
