--- backend/src/main/java/kernel360/techpick/config/JpaAuditingConfig.java
@@ -1,8 +1,0 @@
-package kernel360.techpick.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {}
