+++ backend/src/main/resources/application.yaml
@@ -4,7 +4,7 @@
 
 spring:
   application:
-    name: techpick-server
+    name: linkpick-server
   output:
     ansi:
       enabled: always
