+++ backend/src/test/java/kernel360/linkpick/LinkpickApplicationTests.java
@@ -1,0 +1,13 @@
+package kernel360.linkpick;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LinkpickApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
