--- backend/src/test/java/kernel360/techpick/TechpickApplicationTests.java
@@ -1,13 +1,0 @@
-package kernel360.techpick;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TechpickApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
